Moving simpletestcases.
